split tokens
